requirements.txt
# requirements.txt
apache-airflow-providers-postgres
pandas
sqlalchemy
psycopg2-binary